requirements.txt
pydantic
python-dotenv
pymupdf
numpy
pillow
requests
pydantic-ai
flask
flask-session
werkzeug
apscheduler
gunicorn
flask
dotenv
python-docx
requests